- 로또 랜덤 번호 생성 완료 - DB에 로또 info : insert 처리 완료

+++ LoootCreate/Models/Lottery2026Info.cs
@@ -57,7 +57,7 @@ public class Lottery2026Data
 
 public class LotteryNumber
 {
-    public byte lotteryid { get; set; }
+    public int lotteryid { get; set; }
     public byte Num1 { get; set; }
     public byte Num2 { get; set; }
     public byte Num3 { get; set; }
